Trait search_trees::base::QueryableTreeNode

pub trait QueryableTreeNode<T: Ord + Copy + Debug> {
    fn get_left(&self) -> &Option<Rc<RefCell<Self>>>;

    fn get_right(&self) -> &Option<Rc<RefCell<Self>>>;

    fn get_data(&self) -> T;

    fn height(&self) -> usize { ... }

    fn count_leaves(&self) -> usize { ... }

    fn print_inorder(&self) { ... }

    fn min(&self) -> T { ... }

    fn max(&self) -> T { ... }

    fn contains(&self, value: T) -> bool { ... }

    fn len(&self) -> usize { ... }
}

Provide query functions for nodes

Required methods

fn get_left(&self) -> &Option<Rc<RefCell<Self>>>

Get left child node

fn get_right(&self) -> &Option<Rc<RefCell<Self>>>

Get right child node

fn get_data(&self) -> T

Get data from current node

Loading content...
Provided methods

fn height(&self) -> usize

Return the height of current node, which will be called by QueryableTree.height

fn count_leaves(&self) -> usize

Return the number of leaves, which will be called by QueryableTree.count_leaves

fn print_inorder(&self)

Print nodes inorder , which will be called by QueryableTree.print_inorder

fn min(&self) -> T

Return the minimum value of current node, which will be called by QueryableTree.min

fn max(&self) -> T

Return the maximum value of current node, which will be called by QueryableTree.max

fn contains(&self, value: T) -> bool

Determine whether the node and its successors contains given value, which will be called by QueryableTree.contains

fn len(&self) -> usize

Return the length of the current node, which will be called by QueryableTree.len

Loading content...